Crate compose_error_codes

Structs

ErrorCode

Constants

E0001_UNCLOSED_DELIMITER

E0001: Unclosed delimiter

E0002_INVALID_ASSIGNMENT

E0002: Assignments are not allowed in expression contexts

E0003_EXPECTED_BINDING_AFTER_LET

E0003: Expected binding after let

E0004_MUTATE_IMMUTABLE_VARIABLE

E0004: Cannot reassign to a variable declared as immutable

E0005_IF_EXPRESSION_BODIES_REQUIRE_BRACES

E0005: if expressions must use braces to delimit their body.

E0006_UNTERMINATED_STATEMENT

E0006: A semicolon is required after each statement.

E0007_MISSING_EQUALS_AFTER_LET_BINDING

E0007: A let binding must be followed by = or terminated with a semicolon or newline.

E0008_EXPECTED_EXPRESSION

E0008: An expression was expected but none was found

E0009_ARGS_MISSING_COMMAS

E0009: Function arguments must be separated by commas

E0010_UNCAPTURED_VARIABLE

E0010: Outer variables used in closure but not captured

E0011_UNBOUND_VARIABLE

E0010: Unbound variable

E0012_PREDICATE_MUST_RETURN_BOOLEAN

E0012: Predicate functions must return a boolean value

W0001_USED_UNINITIALIZED_VARIABLE

W0001: Use of a variable before it has been initialized

Functions

lookup

Find an error code by its code.